cambios varios

+++ database/migrations/2019_07_10_164712_create_shopping_table.php.php
@@ -16,12 +16,10 @@ class CreateShoppingTable extends Migration
         Schema::create('shopping', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('product_id');
-
-
-          $table->foreign('user_id')->references('id')->on('users');
-          $table->foreign('product_id')->references('id')->on('products');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('product_id');
+            $table->foreign('user_id')->references('id')->on('users');
+            $table->foreign('product_id')->references('id')->on('products');
         });
     }
